Move sonicd and sonictool implementations into libraries to faciliate reuse
