+++ cmd/sonicd/app/app.go
@@ -1,0 +1,11 @@
+package app
+
+import (
+	"os"
+)
+
+func Run() error {
+	initApp()
+	initAppHelp()
+	return app.Run(os.Args)
+}
