+++ cmd/sonicd/app/testdata/dupes/foo
