+++ cmd/sonicd/app/testdata/empty.js
