+++ cmd/sonicd/app/testdata/guswallet.json
