+++ cmd/sonicd/app/testdata/keystore/.hiddenfile
