+++ cmd/sonicd/app/testdata/keystore/README
