+++ cmd/sonicd/app/testdata/keystore/aaa
