+++ cmd/sonicd/app/testdata/keystore/foo/fd9bd350f08ee3c0c19b85a8e16114a11a60aa4e
