+++ cmd/sonicd/app/testdata/keystore/garbage
