+++ cmd/sonicd/app/testdata/keystore/no-address
