+++ cmd/sonicd/app/testdata/keystore/zero
