+++ cmd/sonicd/app/testdata/keystore/zzz
