+++ cmd/sonicd/app/testdata/passwords.txt
