+++ cmd/sonicd/app/testdata/wrong-passwords.txt
