+++ cmd/sonictool/app/chain.go
@@ -1,4 +1,4 @@
-package main
+package app
 
 import (
 	"compress/gzip"
