+++ cmd/sonictool/app/cli.go
@@ -1,4 +1,4 @@
-package main
+package app
 
 import (
 	"fmt"
