+++ cmd/sonictool/app/export_genesis.go
@@ -1,4 +1,4 @@
-package main
+package app
 
 import (
 	"context"
